requirements.txt
beautifulsoup4
matplotlib
numpy
python-slugify
requests
scipy
